+++ CHANGES.md
@@ -1,9 +1,12 @@
 # Changes
 
-## 7.1.4 (...)
+## 7.2.0 (...)
 - always use PKCE
 - use 256 bits random data for "state"
-- simplify `Random` interface
+- simplify `RandomInterface` and implementation
+- use separate JSON class
+- remove some "setters" from `OAuthClient` class as they were only needed for
+  testing
 
 ## 7.1.3 (2018-06-02)
 - no need to use `strcasecmp`